fix(audit): pass 1 — split windowFocused out of the over-ceiling lifecycle TU
engine_lifecycle.cpp is a grandfathered over-ceiling file (1288 lines on
main) and this PR grew it further, which the file-size rule counts as a
finding. Move windowFocused verbatim into engine_focus.cpp, the same split
this TU has taken on each of its four earlier ceiling crossings, leaving
engine_lifecycle.cpp at 1084 lines.

Also drop a redundant layoutParamsForScreen re-resolve in
cancelDragInsertPreview's same-key view restore: the identical params value
is already in scope.
