+++ libs/phosphor-scroll-engine/CMakeLists.txt
@@ -46,6 +46,7 @@ add_library(PhosphorScrollEngine SHARED
     src/scrollengine/drag_autoscroll.cpp
     src/scrollengine/drag_preview.cpp
     src/scrollengine/engine_lifecycle.cpp
+    src/scrollengine/engine_focus.cpp
     src/scrollengine/engine_closehold.cpp
     src/scrollengine/engine_minsize.cpp
     src/scrollengine/engine_reopen.cpp
